+++ packages/backend/src/apps/custom-api/__tests__/common/size-monitor.test.ts
@@ -1,7 +1,7 @@
 import { Readable, Transform, Writable } from 'stream'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 
-import { createSizeMonitor } from '@/helpers/http-client/size-monitor'
+import { createSizeMonitor } from '@/apps/custom-api/common/size-monitor'
 
 const mocks = vi.hoisted(() => ({
   warn: vi.fn(),
