+++ packages/backend/src/apps/custom-api/common/stream-response.ts
